# src/api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Multi-Chain Smart Contract Analytics API
  description: |
    Comprehensive blockchain analytics platform that analyzes smart contracts across multiple chains 
    with intelligent chain isolation and enterprise-grade reliability.
    
    ## Features
    - **Multi-Chain Support**: Ethereum, Lisk, Starknet, Polygon, Base, Arbitrum, Optimism
    - **OAuth Authentication**: Google and GitHub OAuth integration
    - **Multiple Contract Configurations**: Save and manage multiple analysis configurations
    - **Analysis History**: Track all your analyses with detailed results
    - **Comprehensive Analytics**: 20+ DeFi metrics across user lifecycle, activity, financial, and performance categories
    - **Multiple Output Formats**: JSON, CSV, and Markdown reports
    
    ## Authentication
    This API supports two authentication methods:
    1. **JWT Bearer Token**: Use the `/api/auth/login` endpoint to get a token
    2. **API Key**: Use the `X-API-Key` header with your personal API key
    
    ## Rate Limits
    - General endpoints: 100 requests per 15 minutes
    - Analysis endpoints: 10 requests per hour
    - Limits vary by user tier (Free, Pro, Enterprise)
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@analytics-api.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.analytics-platform.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Contract:
      type: object
      required:
        - address
        - chain
        - name
      properties:
        address:
          type: string
          example: "0x1234567890123456789012345678901234567890"
        chain:
          type: string
          enum: [ethereum, lisk, starknet, polygon, base, arbitrum, optimism]
          example: "ethereum"
        name:
          type: string
          example: "USDT Token"
        abi:
          type: string
          nullable: true
          example: "./abis/usdt.json"
    ContractConfig:
      type: object
      required:
        - name
        - targetContract
        - rpcConfig
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        userId:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "USDT Analysis Config"
        description:
          type: string
          example: "Analysis configuration for USDT token on Ethereum"
        targetContract:
          $ref: '#/components/schemas/Contract'
        competitors:
          type: array
          items:
            $ref: '#/components/schemas/Contract'
        rpcConfig:
          type: object
          properties:
            ethereum:
              type: array
              items:
                type: string
              example: ["https://eth.nownodes.io/api-key"]
            lisk:
              type: array
              items:
                type: string
              example: ["https://lisk.drpc.org"]
        analysisParams:
          type: object
          properties:
            blockRange:
              type: integer
              minimum: 100
              maximum: 10000
              default: 1000
            whaleThreshold:
              type: number
              default: 10
            maxConcurrentRequests:
              type: integer
              minimum: 1
              maximum: 20
              default: 5
            failoverTimeout:
              type: integer
              default: 30000
            maxRetries:
              type: integer
              minimum: 1
              maximum: 5
              default: 2
            outputFormats:
              type: array
              items:
                type: string
                enum: [json, csv, markdown]
        tags:
          type: array
          items:
            type: string
        isActive:
          type: boolean
          default: true
        lastAnalyzed:
          type: string
          format: date-time
          nullable: true
        analysisCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: array
          items:
            type: string
          nullable: true
    PaginationResponse:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 25
        pages:
          type: integer
          example: 3
paths:
...
  /api/contracts:
    get:
      summary: Get user's contract configurations
      description: Retrieve all contract configurations for the authenticated user
      tags: [Contracts]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of items per page
        - in: query
          name: search
          schema:
            type: string
          description: Search term for filtering configurations
        - in: query
          name: chain
          schema:
            type: string
            enum: [ethereum, lisk, starknet, polygon, base, arbitrum, optimism]
          description: Filter by blockchain network
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
          description: Filter by tags
      responses:
        '200':
          description: Contract configurations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContractConfig'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'

    post:
      summary: Create new contract configuration
      description: Create a new contract configuration for analysis
      tags: [Contracts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - targetContract
                - rpcConfig
              properties:
                name:
                  type: string
                  example: "USDT Analysis"
                description:
                  type: string
                  example: "Analysis configuration for USDT token"
                targetContract:
                  $ref: '#/components/schemas/Contract'
                competitors:
                  type: array
                  items:
                    $ref: '#/components/schemas/Contract'
                rpcConfig:
                  type: object
                  example:
                    ethereum: ["https://eth.nownodes.io/api-key"]
                    lisk: ["https://lisk.drpc.org"]
                analysisParams:
                  type: object
                  properties:
                    blockRange:
                      type: integer
                      default: 1000
                    whaleThreshold:
                      type: number
                      default: 10
                tags:
                  type: array
                  items:
                    type: string
                  example: ["defi", "token"]
      responses:
        '201':
          description: Contract configuration created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Contract configuration created successfully"
                  config:
                    $ref: '#/components/schemas/ContractConfig'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...